pumpkin/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file src/requirements.txt
alembic==1.14.1
    # via auth-sprint-1 (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-fastapi-jwt-auth==0.6.6
    # via auth-sprint-1 (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==44.0.1
    # via python-jose
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via fastapi
fastapi==0.115.4
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   async-fastapi-jwt-auth
fastapi-cli==0.0.7
    # via fastapi
fastapi-pagination==0.12.34
    # via auth-sprint-1 (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   async-fastapi-jwt-auth
    #   fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via fastapi
mako==1.3.9
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.15
    # via auth-sprint-1 (pyproject.toml)
packaging==24.2
    # via pytest
passlib==1.7.4
    # via auth-sprint-1 (pyproject.toml)
pluggy==1.5.0
    # via pytest
psycopg==3.2.5
    # via auth-sprint-1 (pyproject.toml)
psycopg-binary==3.2.5
    # via auth-sprint-1 (pyproject.toml)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   fastapi
    #   fastapi-pagination
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.0
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   async-fastapi-jwt-auth
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via async-fastapi-jwt-auth
pytest==8.3.5
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==0.25.3
    # via auth-sprint-1 (pyproject.toml)
pytest-mock==3.14.0
    # via auth-sprint-1 (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.4.0
    # via auth-sprint-1 (pyproject.toml)
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
redis==5.2.0
    # via auth-sprint-1 (pyproject.toml)
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
rsa==4.9
    # via python-jose
setuptools==75.8.0
    # via auth-sprint-1 (pyproject.toml)
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   alembic
starlette==0.41.3
    # via fastapi
typer==0.15.1
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   fastapi-pagination
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
uvicorn==0.34.0
    # via
    #   auth-sprint-1 (pyproject.toml)
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0
    # via uvicorn
werkzeug==3.1.3
    # via auth-sprint-1 (pyproject.toml)
